move chan_open_init
