+++ crates/ibc/src/core/context.rs
@@ -2,6 +2,7 @@ mod chan_close_confirm;
 mod chan_close_init;
 mod chan_open_ack;
 mod chan_open_confirm;
+mod chan_open_init;
 mod chan_open_try;
 
 mod acknowledgement;
@@ -14,6 +15,7 @@ use self::chan_close_confirm::{chan_close_confirm_execute, chan_close_confirm_va
 use self::chan_close_init::{chan_close_init_execute, chan_close_init_validate};
 use self::chan_open_ack::{chan_open_ack_execute, chan_open_ack_validate};
 use self::chan_open_confirm::{chan_open_confirm_execute, chan_open_confirm_validate};
+use self::chan_open_init::{chan_open_init_execute, chan_open_init_validate};
 use self::chan_open_try::{chan_open_try_execute, chan_open_try_validate};
 
 use self::acknowledgement::{acknowledgement_packet_execute, acknowledgement_packet_validate};
@@ -36,13 +38,10 @@ use crate::core::ics03_connection::connection::ConnectionEnd;
 use crate::core::ics03_connection::version::{
     get_compatible_versions, pick_version, Version as ConnectionVersion,
 };
-use crate::core::ics04_channel::channel::{ChannelEnd, Counterparty, State};
+use crate::core::ics04_channel::channel::ChannelEnd;
 use crate::core::ics04_channel::commitment::{AcknowledgementCommitment, PacketCommitment};
 use crate::core::ics04_channel::context::calculate_block_delay;
-use crate::core::ics04_channel::events::OpenInit;
-use crate::core::ics04_channel::handler::chan_open_init;
 use crate::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
-use crate::core::ics04_channel::msgs::chan_open_init::MsgChannelOpenInit;
 use crate::core::ics04_channel::msgs::{ChannelMsg, PacketMsg};
 use crate::core::ics04_channel::packet::{Receipt, Sequence};
 use crate::core::ics04_channel::timeout::TimeoutHeight;
@@ -622,103 +621,4 @@ pub trait ExecutionContext: ValidationContext {
 
     /// Logging facility
     fn log_message(&mut self, message: String);
-}
-
-fn chan_open_init_validate<ValCtx>(
-    ctx_a: &ValCtx,
-    module_id: ModuleId,
-    msg: MsgChannelOpenInit,
-) -> Result<(), ContextError>
-where
-    ValCtx: ValidationContext,
-{
-    chan_open_init::validate(ctx_a, &msg)?;
-    let chan_id_on_a = ChannelId::new(ctx_a.channel_counter()?);
-
-    let module = ctx_a
-        .get_route(&module_id)
-        .ok_or(ChannelError::RouteNotFound)?;
-    module.on_chan_open_init_validate(
-        msg.ordering,
-        &msg.connection_hops_on_a,
-        &msg.port_id_on_a,
-        &chan_id_on_a,
-        &Counterparty::new(msg.port_id_on_b.clone(), None),
-        &msg.version_proposal,
-    )?;
-
-    Ok(())
-}
-
-fn chan_open_init_execute<ExecCtx>(
-    ctx_a: &mut ExecCtx,
-    module_id: ModuleId,
-    msg: MsgChannelOpenInit,
-) -> Result<(), ContextError>
-where
-    ExecCtx: ExecutionContext,
-{
-    let chan_id_on_a = ChannelId::new(ctx_a.channel_counter()?);
-    let module = ctx_a
-        .get_route_mut(&module_id)
-        .ok_or(ChannelError::RouteNotFound)?;
-    let (extras, version) = module.on_chan_open_init_execute(
-        msg.ordering,
-        &msg.connection_hops_on_a,
-        &msg.port_id_on_a,
-        &chan_id_on_a,
-        &Counterparty::new(msg.port_id_on_b.clone(), None),
-        &msg.version_proposal,
-    )?;
-
-    let conn_id_on_a = msg.connection_hops_on_a[0].clone();
-
-    // state changes
-    {
-        let port_channel_id_on_a = (msg.port_id_on_a.clone(), chan_id_on_a.clone());
-        let chan_end_on_a = ChannelEnd::new(
-            State::Init,
-            msg.ordering,
-            Counterparty::new(msg.port_id_on_b.clone(), None),
-            msg.connection_hops_on_a.clone(),
-            msg.version_proposal.clone(),
-        );
-
-        ctx_a.store_channel(port_channel_id_on_a.clone(), chan_end_on_a)?;
-
-        ctx_a.increase_channel_counter();
-
-        // Associate also the channel end to its connection.
-        ctx_a.store_connection_channels(conn_id_on_a.clone(), port_channel_id_on_a.clone())?;
-
-        // Initialize send, recv, and ack sequence numbers.
-        ctx_a.store_next_sequence_send(port_channel_id_on_a.clone(), 1.into())?;
-        ctx_a.store_next_sequence_recv(port_channel_id_on_a.clone(), 1.into())?;
-        ctx_a.store_next_sequence_ack(port_channel_id_on_a, 1.into())?;
-    }
-
-    // emit events and logs
-    {
-        ctx_a.log_message(format!(
-            "success: channel open init with channel identifier: {chan_id_on_a}"
-        ));
-        let core_event = IbcEvent::OpenInitChannel(OpenInit::new(
-            msg.port_id_on_a.clone(),
-            chan_id_on_a.clone(),
-            msg.port_id_on_b,
-            conn_id_on_a,
-            version,
-        ));
-        ctx_a.emit_ibc_event(core_event);
-
-        for module_event in extras.events {
-            ctx_a.emit_ibc_event(IbcEvent::AppModule(module_event));
-        }
-
-        for log_message in extras.log {
-            ctx_a.log_message(log_message);
-        }
-    }
-
-    Ok(())
 }
